# apps/backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aspiron API",
    "description": "Backend API for the Aspiron project",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/api/v1/admin/insights": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Insights"
        ],
        "operationId": "handler_get_insights",
        "parameters": [
          {
            "name": "pagination",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationPayload"
            }
          },
          {
            "name": "filter",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InsightFilters"
            }
          },
          {
            "name": "sort",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/InsightSort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get admin insights with pagination and filtering"
          }
        }
      }
    },
    "/api/v1/admin/insights/topics": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Insights"
        ],
        "operationId": "handler_get_topic_performance",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "pagination",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationPayload"
            }
          },
          {
            "name": "sort",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TopicPerformanceSort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get topic performance insights"
          },
          "401": {
            "description": "Unauthorized - Missing VIEW_ANALYTICS permission"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/chapters/{chapter_id}/topics": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "handler_get_topics_by_chapter_id",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get topics by chapter ID"
          }
        }
      }
    },
    "/api/v1/community/threads": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Community"
        ],
        "operationId": "handler_create_thread",
        "responses": {
          "200": {
            "description": "Create a community thread"
          }
        }
      }
    },
    "/api/v1/community/threads/public": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Community"
        ],
        "operationId": "handler_public_threads",
        "responses": {
          "200": {
            "description": "Get public threads"
          }
        }
      }
    },
    "/api/v1/community/threads/{thread_id}": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Community"
        ],
        "operationId": "handler_get_thread_by_id",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "description": "Thread ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch community thread by ID"
          }
        }
      }
    },
    "/api/v1/community/threads/{thread_id}/attach-note": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Community"
        ],
        "operationId": "handler_attach_note",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "description": "Thread ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attach note to thread"
          }
        }
      }
    },
    "/api/v1/community/threads/{thread_id}/posts": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Community"
        ],
        "operationId": "handler_create_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "description": "Thread ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Create community posts"
          }
        }
      }
    },
    "/api/v1/community/topics/{topic_id}/threads": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Community"
        ],
        "operationId": "handler_get_threads_by_topic",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch community threads per topic"
          }
        }
      }
    },
...
    "/api/v1/subjects": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "handler_get_all_subjects",
        "responses": {
          "200": {
            "description": "Get all subjects"
          }
        }
      }
    },
    "/api/v1/subjects/{subject_id}/chapters": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "handler_get_chapters_by_subject_id",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "description": "Subject ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get chapters by subject ID"
          }
        }
      }
    },
    "/api/v1/topics/{topic_id}": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "handler_get_topic_by_id",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get topic metadata by topic ID"
          }
        }
      }
    },
...
    "/api/v1/topics/{topic_id}/videos": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "handler_get_videos_by_topic_id",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get videos by topic ID"
          }
        }
      }
    },
...
    "/api/v1/videos/{video_id}/offline-token": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "handler_get_offline_token_by_video_id",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "description": "Video ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get offline token by video ID"
          }
        }
      }
    }
  },
...
}